# doc/swagger.yaml
swagger: "2.0"
info:
  description: ""
  version: "0.1.0"
  title: "KilnRegulator"
  contact:
    url: "https://github.com/jdeantoni/KilnRegulator3.0/issues"
  license:
    name: "MIT"
host: "127.0.0.1:3000"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /programs:
    get:
      tags:
      - "programs"
      summary: "Get all programs"
      operationId: "getPrograms"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "all programs"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Program"
    post:
      tags:
      - "programs"
      summary: "add new program"
      operationId: "addProgram"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Program"
      responses:
        200:
          description: "successful operation"
        409:
          description: "program already exists (uuid not unique)"
        500:
          description: "internal server error (database)"
  /programs/{uuid}:
    get:
      tags:
      - "programs"
      summary: "Get one program"
      operationId: "getProgram"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "uuid"
        type: "string"
        required: true
        format: "uuid"
      responses:
        200:
          description: "all programs"
          schema:
            $ref: "#/definitions/Program"
        400:
          description: "missing uuid"
        404:
          description: "old program not found"
    put:
      tags:
      - "programs"
      summary: "edit program, ie. archive old program if referenced or delete it, and add new program"
      operationId: "editProgram"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "uuid"
        type: "string"
        required: true
        format: "uuid"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Program"
      responses:
        200:
          description: "successful operation"
        400:
          description: "missing uuid"
        404:
          description: "old program not found"
        500:
          description: "internal server error (database)"
    delete:
      tags:
      - "programs"
      summary: "delete program, ie. archive old program if referenced or delete it"
      operationId: "deleteProgram"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "uuid"
        type: "string"
        required: true
        format: "uuid"
      responses:
        200:
          description: "successful operation"
        400:
          description: "missing uuid"
        404:
          description: "old program not found"
        500:
          description: "internal server error (database)"
  /action/start:
    post:
      tags:
      - "action"
      summary: "Start cooking"
      operationId: "startCooking"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          type: "object"
          properties:
            uuid:
              type: "string"
              description: "UUID of program to start on oven"
              format: "uuid"
      - in: "query"
        type: "number"
        name: "delay"
        required: false
        description: "optional time for delayed cooking, in hour"
        format: "number"
      responses:
        200:
          description: "successful operation"
        400:
          description: "invalid operation, ex. already started or program not found"
        500:
          description: "could not start cooking"
...
definitions:
...
  Segment:
    type: "object"
    properties:
      targetTemperature:
        type: "integer"
        description: "target temperature at the end of the segment"
        format: "int64"
      slope:
        type: "number"
        description: "slope in °C/second"
        format: "float"
      duration:
        type: "integer"
        description: "duration in seconds"
        format: "int64"
  Program:
    type: "object"
    properties:
      uuid:
        type: "string"
        description: "Program UUID used for this cooking"
        format: "uuid"
      name:
        type: "string"
        description: "Program name"
      lastModificationDate:
        type: "string"
        description: "Date and time when this program was last modified"
        format: "date-time"
      segments:
        type: "array"
        items:
          $ref: "#/definitions/Segment"
      segmentEditableStates:
        type: "array"
        items:
          type: "boolean"
...